refactor: extract entry point service logic into `Stadium`
